# Performance: Fix empty renders of LibraryMoveable