--- a/packages/story-editor/src/app/pageCanvas/getPageWithoutSelection.js
+++ b/packages/story-editor/src/app/pageCanvas/getPageWithoutSelection.js
@@ -16,7 +16,7 @@
 function pageWithoutSelection(page, selection) {
   return {
     ...page,
-    elements: page.element.filter(({ id }) => !selection.includes(id)),
+    elements: page.elements.filter(({ id }) => !selection.includes(id)),
   };
 }