## Context
Noticed while doing some performance work that we had different performance metrics for the same actions depending on what library panel was open. Specifically the shapes panel and the text panel around selection. Looked into it and found out it was our `LibraryMoveable` component. Theorized some possible approaches and the exploration in this ticket #10292.

Ultimately created this ticket #10730 as the best approach to fix the empty re-renders in the library panels. 

Also saw a possible win with the new approach to centralize some of our `html-to-canvas` usage. This sets up the foundation for that and hopefully we can share more generated canvases from story pages amongst components in the editor that need it. Created these follow tickets to look into that. Measured how long our `htmlToImage.toCanvas()` call took and on our test story with 50 shape elements it took ~1300ms, so any way we can minimize this call is a big win:

* #10822
* #10821
* #10824

Also with this new architecture we were able to eliminate some complexity in the logic, passing around of refs to a provider, and having to `JSON.stringify()` the entire story page every time it updated 🙌 . Relevant prior art here #10054

## Summary
Restructures how our smart colors are generating and passing around canvases to read color data off of them. Has a couple benefits that I'll go over more in the technical choices, but the big performance win here is limiting empty re-renders on selection when the text or shape panel are open.

| Selection update, text panel open | Selection update, shapes panel open |
| ---- | ----- | 
|  <img width="1440" alt="Screen Shot 2022-03-08 at 3 10 15 PM" src="https://user-images.githubusercontent.com/35983235/157325604-550d9c21-0efe-4f48-8262-437f344874a4.png"> |  <img width="1440" alt="Screen Shot 2022-03-08 at 3 10 05 PM" src="https://user-images.githubusercontent.com/35983235/157325629-39b2f173-ca7c-4ce1-8135-f416b02d6484.png"> |



**Selection update, text panel open**
| Before | After |
| ----- | ----- |
| react commit  | react commit |
| ~46ms | ~26ms |
| <img width="585" alt="Screen Shot 2022-03-08 at 2 47 01 PM" src="https://user-images.githubusercontent.com/35983235/157322504-a9a70fd2-4cd4-4849-804f-2b12de861698.png"> |  <img width="589" alt="Screen Shot 2022-03-08 at 2 48 01 PM" src="https://user-images.githubusercontent.com/35983235/157322516-1fd3fdfa-8463-446d-9c0d-40145117a433.png"> |
| total js | total js |
| ~103ms | ~82ms |
| <img width="588" alt="Screen Shot 2022-03-08 at 2 54 05 PM" src="https://user-images.githubusercontent.com/35983235/157323459-1134c04d-a83b-4d8c-8046-787a5b3c47eb.png"> |  <img width="590" alt="Screen Shot 2022-03-08 at 2 55 06 PM" src="https://user-images.githubusercontent.com/35983235/157323484-0630415a-92b8-47c8-97c0-0dc70c226ae8.png"> |

**Selection update, shape panel open**
| Before | After |
| ----- | ----- |
| react commit  | react commit |
| ~86ms | ~26ms |
| <img width="630" alt="Screen Shot 2022-03-08 at 3 01 54 PM" src="https://user-images.githubusercontent.com/35983235/157324740-6392197a-8752-4fac-b2f0-d841dcbe040f.png"> |  <img width="590" alt="Screen Shot 2022-03-08 at 3 02 18 PM" src="https://user-images.githubusercontent.com/35983235/157324759-3342c687-e3b9-47bc-9abe-55b54616427b.png"> |
| total js | total js |
| ~145ms | ~87ms |
|  <img width="632" alt="Screen Shot 2022-03-08 at 3 07 18 PM" src="https://user-images.githubusercontent.com/35983235/157325286-1333ced8-ac00-422f-b084-3349bb5a0f6e.png"> |  <img width="638" alt="Screen Shot 2022-03-08 at 3 07 28 PM" src="https://user-images.githubusercontent.com/35983235/157325304-a98e54f7-5cd0-4658-b376-0a064661dfd9.png"> |


## Relevant Technical Choices
- Central provider responsible for holding and invalidating page canvas caches `PageCanvasProvider` 
- Creates & uses `storyPageToCanvas` util which simplifies some of the logic by keeping the current state of the user facing story canvas ui state agnostic from generating an image with the story markup.
- Makes `calculateAccessibleTextColors` stable throughout the life of the application
- Extracts & creates `useIdleQueue` from `pageUrlProvider`
- Simplifies cache invalidation logic to use react state vs storing & looking at `pageHash`
- Removes `pageHash`
- Removes `usePageAsCanvas`
- Cleans up some now unneeded state in the `CanvasProvider`

## To-do

<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes
NA
<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions
Go through and see that text smart colors behave the same as on `main`:
<img width="1439" alt="Screen Shot 2022-03-08 at 3 32 04 PM" src="https://user-images.githubusercontent.com/35983235/157328883-9b365c73-26c4-4e62-aec2-b2124d92c9dc.png">

<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [x] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

**Adaptive color action**

1. add text on poor contrast backgrounds
2. select text
3. click on  `adaptive colors` action in the design panel with the paint bucket (highlighted above)
4. see that text becomes better contrast on the background

**Adaptive color toggle**

1. set the background color to black
2. toggle on `Adaptive Text Colors` in the library (highlighted above)
3. add a preset (like `Title 1`) and see that it gets applied to the page white
4. repeat on a white background and see that text is black


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/googleforcreators/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/googleforcreators/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10730