@littlemilkstudio 2022-03-14
> There are some tests related to smart color (e.g. should add text color based on background) which have a delay of 800 ms currently to wait for canvas generation + color generation. Wondering if those delays could be reduced now?

I don't think there's any speed increase here. all the changes relative to this were just for code quality. We're still generating a new canvas for both the actions in the specified karma test, so just gonna leave it as is at 800 for now