@swissspidy 2022-03-16
Looking at the Percy snapshots now...

Looks like the backdrop from the media edit mode now goes all the way over the carousel? Is that new?

<img width="1711" alt="Screenshot 2022-03-16 at 09 50 47" src="https://user-images.githubusercontent.com/841956/158552161-97b7a7e7-3d07-415e-8d96-de33564f3160.png">

See https://percy.io/google/web-stories-wp/builds/16492104/changed/928097483