@miina 2022-03-16
> Looking at the Percy snapshots now...
> Looks like the backdrop from the media edit mode now goes all the way over the carousel? Is that new?

@swissspidy I think #10929 will fix it.